Update README.md
Updating the read me. There still needs to be some improvement. I'm sure there are spelling and grammar errors that I missed that need to be corrected. I'm posting this here because it is better for there be some documentation instead of no documentation. I also need to explain what the struct is doing and what data it represents because the original code doesn't really specify, and I don't want to assume. So I will have to run some test and do some investigating. I would also like to add more content to the usage section to explain why and how to use the newly added functions, even though this is sort of documented in the INA219.c file.